+++ R/utils.R
@@ -92,3 +92,12 @@ load_dir_to_temp <- function(n = 4) {
 
   return(dest_dir)
 }
+
+download_to_temp <- function(url) {
+  url <- URLencode(url)
+  extt <- tools::file_ext(url)
+  tmpfi <- tempfile(fileext = paste0(".", extt))
+  rq <- httr2::request(url)
+  dwn <- httr2::req_perform(rq, path = tmpfi)
+  return(tmpfi)
+}