+++ NEWS.md
@@ -1,5 +1,6 @@
 # resmush 0.1.0
 
+-   Now **resmush** uses **httr2** instead of **httr**.
 -   Add DOI: <https://doi.org/10.5281/zenodo.10556679>.
 -   `resmush_file()`, `resmush_dir()` and `resmush_url()` gain an new argument
     `overwrite`.